Refactor: StoreItem 컴포넌트화
