+++ src/components/main/index.js
@@ -3,4 +3,8 @@ export { default as StoreItemOnHover } from './StoreItemOnHover';
 export { default as NoResultMessage } from './NoResultMessage';
 export { default as InfiniteStoreList } from './InfiniteStoreList';
 export { default as MychelinStars } from './MychelinStars';
+export { default as TopThreeStores } from './TopThreeStores';
+export { default as UnrankedStores } from './UnrankedStores';
+export { default as StoreItemOnNoHover } from './StoreItemOnNoHover';
+export { default as StoreItemOnHOver } from './StoreItemOnHover';
 export { default as StoreItem } from './StoreItem';
